+++ src/main/java/ddingdong/ddingdongBE/domain/scorehistory/entity/Score.java
@@ -1,5 +1,6 @@
 package ddingdong.ddingdongBE.domain.scorehistory.entity;
 
+import java.math.BigDecimal;
 import java.util.Objects;
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
@@ -15,9 +16,9 @@ import lombok.NoArgsConstructor;
 public class Score {
 
     @Column(name = "score", nullable = false, columnDefinition = "DECIMAL(10,3) DEFAULT 0")
-    private float value;
+    private BigDecimal value;
 
-    private Score(float value) {
+    private Score(BigDecimal value) {
         this.value = value;
     }
 
@@ -38,9 +39,8 @@ public class Score {
         return Objects.hash(getValue());
     }
 
-    public static Score from(float value) {
+    public static Score from(BigDecimal value) {
         return new Score(value);
     }
-
 
 }
